+++ src/main/java/cl/catastrofescl/citizen/exception/DonationQuantityExceededException.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 public class DonationQuantityExceededException extends RuntimeException {
 
-    public DonationQuantityExceededException(UUID itemId, int cantidadSolicitada, int cantidadMaxima) {
+    public DonationQuantityExceededException(UUID itemId, long cantidadSolicitada, long cantidadMaxima) {
         super(String.format(
                 "La cantidad solicitada (%d) para el ítem %s supera el cupo disponible (%d). "
                         + "El centro ya tiene suficiente stock comprometido de ese recurso.",
